Update docs.
Also move config to its own file.
